requirements.txt
qrcode==7.3.1
Flask==2.2.2
Pillow==9.3.0
hubspain==1.0.1
python-dotenv==0.21.0